initialize react-gtm module
