+++ frontend/src/app/layout.tsx
@@ -1,23 +1,6 @@
 import type { Metadata } from 'next'
-import { Geist, Geist_Mono } from 'next/font/google'
 import React from 'react'
-import { Providers } from 'wrappers/provider'
-import BreadCrumbs from 'components/BreadCrumbs'
-import Footer from 'components/Footer'
-
-import Header from 'components/Header'
-import './globals.css'
-import ScrollToTop from 'components/ScrollToTop'
-
-const geistSans = Geist({
-  variable: '--font-geist-sans',
-  subsets: ['latin'],
-})
-
-const geistMono = Geist_Mono({
-  variable: '--font-geist-mono',
-  subsets: ['latin'],
-})
+import ClientLayout from './clientLayout'
 
 export const metadata: Metadata = {
   title: 'OWASP Nest',
@@ -53,25 +36,6 @@ export const metadata: Metadata = {
   },
 }
 
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode
-}>) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-        style={{ minHeight: '100vh' }}
-      >
-        <Providers>
-          <Header />
-          <BreadCrumbs />
-          {children}
-          <Footer />
-          <ScrollToTop />
-        </Providers>
-      </body>
-    </html>
-  )
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+  return <ClientLayout>{children}</ClientLayout>
 }
