+++ frontend/src/components/MultiSearch.tsx
@@ -12,6 +12,7 @@ import { debounce } from 'lodash'
 import { useRouter } from 'next/navigation'
 import type React from 'react'
 import { useState, useEffect, useMemo, useCallback, useRef } from 'react'
+import TagManager from 'react-gtm-module'
 import { fetchAlgoliaData } from 'server/fetchAlgoliaData'
 import { ChapterTypeAlgolia } from 'types/chapter'
 import { EventType } from 'types/event'
@@ -78,7 +79,15 @@ const MultiSearchBar: React.FC<MultiSearchBarProps> = ({
       }, 300),
     [eventData, indexes]
   )
-
+  useEffect(() => {
+    TagManager.dataLayer({
+      dataLayer: {
+        event: 'search',
+        search_term: searchQuery,
+        page_path: window.location.pathname,
+      },
+    })
+  }, [searchQuery])
   useEffect(() => {
     return () => {
       debouncedSearch.cancel()
