+++ frontend/src/utils/credentials.ts
@@ -2,9 +2,7 @@ export const API_URL = process.env.NEXT_PUBLIC_API_URL
 export const CSRF_URL = process.env.NEXT_PUBLIC_CSRF_URL
 export const ENVIRONMENT = process.env.NEXT_PUBLIC_ENVIRONMENT
 export const GRAPHQL_URL = process.env.NEXT_PUBLIC_GRAPHQL_URL
-export const GTM_AUTH = process.env.NEXT_PUBLIC_GTM_AUTH
 export const GTM_ID = process.env.NEXT_PUBLIC_GTM_ID
-export const GTM_PREVIEW = process.env.NEXT_PUBLIC_GTM_PREVIEW
 export const IDX_URL = process.env.NEXT_PUBLIC_IDX_URL
 export const RELEASE_VERSION = process.env.NEXT_PUBLIC_RELEASE_VERSION
 export const SENTRY_DSN = process.env.NEXT_PUBLIC_SENTRY_DSN
